+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Various import optimizations
 - Moved `harvest.json` to `harvest.yaml`
 - Removed `config.py`
+- Implemented logging configuration
 
 # 0.2.5
 - Updated to conform with CloudHarvestCorePluginManager 0.2.4